auto-call-agent/pyproject.toml
[project]
name = "auto-call-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-decouple>=3.8",
    "twilio>=9.8.5",
]
